Update README.md
No matter if the Word has more than One definition, The thesaurus will provide you with all possible definitions of the Word. Incase, you mispelled a Word, the Smart-Thesaurus provides the similar words that may be the word you were actually looking for. It runs an algorithm for choosing which words to suggest to the User so that only those words will be suggested which have 0.8 Ratio Similarity to the Original Word.